signature verify complete for alice and bob

+++ src/main/java/org/ncl/cloudcomputing/ttp/TTP.java
@@ -4,8 +4,10 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.security.InvalidKeyException;
 import java.security.KeyFactory;
+import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.security.PublicKey;
 import java.security.Signature;
@@ -113,6 +115,54 @@ public class TTP extends AWSBase implements Runnable {
 		
 		Logger.log("TTP started");
 	}
+	
+	
+	/**
+	 * Hashes a file
+	 * @param file to hash
+	 * @return byte[]
+	 */
+	private byte[] hashFile(File file) {
+		MessageDigest md;
+		byte[] data;
+		byte[] hash = null;
+		try {
+			md = MessageDigest.getInstance("SHA-256");
+			Path path = file.toPath();
+			data = Files.readAllBytes(path);
+			md.update(data);
+			hash = md.digest();
+		} catch (NoSuchAlgorithmException e1) {
+			e1.printStackTrace();
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		return hash;
+	}
+	
+	
+	
+	private boolean verifySignature(byte[] sigAlice, byte[] docHash, byte[] publicKeyBytes) {
+		PublicKey publicKey;
+		Signature sig;
+		try {
+			publicKey = KeyFactory.getInstance("RSA").generatePublic(new X509EncodedKeySpec(publicKeyBytes));
+			sig = Signature.getInstance("SHA256withRSA");
+			sig.initVerify(publicKey);
+			sig.update(docHash);
+			return sig.verify(sigAlice);
+		} catch (NoSuchAlgorithmException e) {
+			e.printStackTrace();
+		} catch (InvalidKeySpecException e) {
+			e.printStackTrace();
+		} catch (InvalidKeyException e) {
+			e.printStackTrace();
+		} catch (SignatureException e) {
+			e.printStackTrace();
+		}
+		return false;
+	}
+	
 
 	public void run() {
 		while (true) {
@@ -131,45 +181,24 @@ public class TTP extends AWSBase implements Runnable {
 				
 				Logger.logReceiveMessageOnSucceed(messageStatus);
 				
-				PublicKey publicKey;
-				boolean correctSignature = false;
-				Signature sig;
-				
 				if (messageStatus == MessageStatus.Alice_to_TTP.getValue()) {
 					String transactionId = message.getMessageAttributes().get("transaction-id").getStringValue();
 					byte[] sigAlice = message.getMessageAttributes().get("sig-alice").getBinaryValue().array();
+					byte[] docHash = message.getMessageAttributes().get("doc-hash").getBinaryValue().array();
 					byte[] publicKeyAlice = message.getMessageAttributes().get("public-key").getBinaryValue().array();
 					String docKey = message.getMessageAttributes().get("doc-key").getStringValue();
 					String filename = message.getMessageAttributes().get("file-name").getStringValue();
 					
-					String path = this.copyDocumentToLocal(filename, docKey);
-					File file = new File(path);
-					
-					try {
-						publicKey = KeyFactory.getInstance("RSA").generatePublic(new X509EncodedKeySpec(publicKeyAlice));
-						sig = Signature.getInstance("SHA256withRSA");
-						sig.initVerify(publicKey);
-						sig.update(sigAlice);
-					} catch (NoSuchAlgorithmException e) {
-						e.printStackTrace();
-					} catch (InvalidKeySpecException e) {
-						e.printStackTrace();
-					} catch (InvalidKeyException e) {
-						e.printStackTrace();
-					} catch (SignatureException e) {
-						e.printStackTrace();
+					if(verifySignature(sigAlice, docHash, publicKeyAlice)) {
+						TransactionItem item = new TransactionItem(transactionId, sigAlice, docHash, docKey, filename);
+						TransactionItem result = this.transactionRepo.insert(item);
+						
+						if (result != null)
+							this.sendMessageToBob(transactionId, sigAlice);
 					}
-					
-
-					
-					
-					
-					
-					TransactionItem item = new TransactionItem(transactionId, sigAlice, docKey, filename);
-					TransactionItem result = this.transactionRepo.insert(item);
-					
-					if (result != null)
-						this.sendMessageToBob(transactionId, sigAlice);
+					else {
+						// terminate the transaction
+					}
 				}
 				else if (messageStatus == MessageStatus.Bob_to_TTP.getValue()) {
 					String transactionId = message.getMessageAttributes().get("transaction-id").getStringValue();
@@ -178,14 +207,19 @@ public class TTP extends AWSBase implements Runnable {
 					
 					// WE NEED TO KNOW IF SIGBOB SENT BY BOB IS CORRECT
 					/* to do this we need Bob's PublicKey.. could send it in the message or get Bob to transmit it to TTP another way? */
-					
 					TransactionItem transaction = this.transactionRepo.getItemById(transactionId);
 					
-					if (transaction != null) {
-						this.sendDocumentKeyToBob(transactionId, transaction.getDocumentKey(), transaction.getFilename());
-						this.sendMessageToAlice(transactionId, sigBob);
-						this.transactionRepo.delete(transactionId);
+					if(verifySignature(transaction.getSignature(), transaction.getDocumentHash(), publicKeyBob)) {
+						if (transaction != null) {
+							this.sendDocumentKeyToBob(transactionId, transaction.getDocumentKey(), transaction.getFilename());
+							this.sendMessageToAlice(transactionId, sigBob);
+							this.transactionRepo.delete(transactionId);
+						}
 					}
+					else {
+						// terminate the transaction
+					}
+					
 				}
 				
 				this.amazonTTPQueue.deleteMessage(message.getReceiptHandle());
